DELTA versioon 2.3.1.23
Selle versiooniga tulevad kaasa järgmised parandused ja täiendused:
156775 EDHS-456 Otsingu tulemite pealkirjade kuvamisel peab olema piirang 20 tm
153796 SIM087 Puhkuse andmise korralduselt (SIM) võtta ära väli Mall 
153800 SIM092 Töövood bloki täiendused 
153803 SIM095 Failinimede kontroll allkirjastamisel 
153966 SIM099 Süsteemi teadete tekstide muudatused 
152369 SIM111 Siselähetuse taotluse ekraanivormi muudatused 
152370 SIM111 Välislähetuse ekraanivormi muudatused 
152366 SIM111, SIM105 Koolitustaotluse ekraanivormi muudatused 
157723 SIM117, EDHS-464 Saab saata tööülesannet checkbox on kontakti andmetes ja kontaktgruppide halduses puudu 

Välislähetuse taotlusel järje loomisel ei kanta ajutiselt üle dokumendi liigi spetsiifilisi andmeid (lähetused ja taotlejad). See pole viga vaid on teadlikult eemaldatud ajutiste probleemide tõttu dokumendi liigi versioneerimisest tulenevalt.

Uue versiooni paigaldamise järel tuleb üle käia järgnevad klassifikaatorid ja kontrollida, et nende väärtused oleksid kõik täisarvulised ja numbrid (näiteks 100; 50; 20):
* dailyAllowanceRate
* trainingAppDailyAllowanceRate

Infoks: uue versiooniga lisanduvad klassifikaatorid:
* countries
* errandApplicationDomesticCostElement
* errandApplicationDomesticCostManager

Uue versiooniga lisanduvad parameetrid:
* errandOrderAbroadDailyAllowanceSum
* trainingApplicationDailyAllowanceSum


git-svn-id: https://svn.smit/tarkvara/edhs/trunk@101155 5fddf92a-9921-11dd-8d32-b999f448c456

+++ delta/common/source/java/ee/webmedia/alfresco/common/bootstrap/AbstractNodeUpdater.java
@@ -267,13 +267,13 @@ public abstract class AbstractNodeUpdater extends AbstractModuleComponent {
             thisRunCompletedSize += batchList.size();
             batchList = new ArrayList<E>(batchSize);
             long endTime = System.currentTimeMillis();
-            double completedPercent = ((long) completedSize) * 100L / ((double) totalSize);
-            double lastDocsPerSec = ((long) i) * 1000L / ((double) (endTime - startTime));
+            double completedPercent = (completedSize) * 100L / ((double) totalSize);
+            double lastDocsPerSec = (i) * 1000L / ((double) (endTime - startTime));
             long thisRunTotalTime = endTime - thisRunStartTime;
-            double totalDocsPerSec = ((long) thisRunCompletedSize) * 1000L / ((double) thisRunTotalTime);
+            double totalDocsPerSec = (thisRunCompletedSize) * 1000L / ((double) thisRunTotalTime);
             long remainingSize = ((long) totalSize) - ((long) completedSize);
-            long divisor = ((long) thisRunCompletedSize) * 60000L;
-            int etaMinutes = ((int) (remainingSize  *  thisRunTotalTime / divisor)) + 1;
+            long divisor = (thisRunCompletedSize) * 60000L;
+            int etaMinutes = ((int) (remainingSize * thisRunTotalTime / divisor)) + 1;
             int etaHours = 0;
             if (etaMinutes > 59) {
                 etaHours = etaMinutes / 60;
